+++ src/rand/mod.rs
@@ -1,5 +1,6 @@
 //! Pseudo-random element generation.
 
+use rand::RngCore;
 pub use winter_crypto::{DefaultRandomCoin as WinterRandomCoin, RandomCoin, RandomCoinError};
 pub use winter_utils::Randomizable;
 
@@ -11,7 +12,7 @@ pub use rpo::RpoRandomCoin;
 /// Pseudo-random element generator.
 ///
 /// An instance can be used to draw, uniformly at random, base field elements as well as [Word]s.
-pub trait FeltRng {
+pub trait FeltRng: RngCore {
     /// Draw, uniformly at random, a base field element.
     fn draw_element(&mut self) -> Felt;