+++ src/utils/mod.rs
@@ -10,17 +10,13 @@ mod kv_map;
 // RE-EXPORTS
 // ================================================================================================
 
-#[deprecated(since = "0.8.3", note = "You should prefer to import from `alloc::*`")]
-pub use alloc::{format, vec};
-
 pub use winter_utils::{
-    boxed, string, uninit_vector, Box, ByteReader, ByteWriter, Deserializable,
-    DeserializationError, Serializable, SliceReader,
+    uninit_vector, ByteReader, ByteWriter, Deserializable, DeserializationError, Serializable,
+    SliceReader,
 };
 
 pub mod collections {
     pub use super::kv_map::*;
-    pub use winter_utils::collections::*;
 }
 
 // UTILITY FUNCTIONS